Importing 2.0 code from svn

+++ code/nooku/libraries/koowa/database/row/table.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @version		$Id: table.php 987 2011-04-04 20:32:35Z stian $
+ * @version		$Id: table.php 1372 2011-10-11 18:56:47Z stian $
  * @category	Koowa
  * @package     Koowa_Database
  * @subpackage  Row
@@ -20,11 +20,11 @@
 class KDatabaseRowTable extends KDatabaseRowAbstract
 {
 	/**
-	 * Table object or identifier (APP::com.COMPONENT.table.NAME)
+	 * Table object or identifier (com://APP/COMPONENT.table.NAME)
 	 *
 	 * @var	string|object
 	 */
-	protected $_table;
+	protected $_table = false;
 
 	/**
      * Object constructor 
@@ -35,10 +35,15 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	{
 		parent::__construct($config);
 
-		// Set the table indentifier
-		if(isset($config->table)) {
-			$this->setTable($config->table);
-		}
+		$this->_table = $config->table;
+			
+		// Reset the row
+        $this->reset();
+            
+        // Reset the row data
+        if(isset($config->data))  {
+            $this->setData($config->data->toArray(), $this->_new);
+        }
 	}
 
 	/**
@@ -52,83 +57,111 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	protected function _initialize(KConfig $config)
 	{
 		$config->append(array(
-			'table'	=> null
+			'table'	=> $this->getIdentifier()->name
 		));
 
 		parent::_initialize($config);
 	}
 
 	/**
-	 * Get the identifier for the table with the same name
-	 *
-	 * @return	KIdentifierInterface
-	 */
-	public function getTable()
-	{
-		if(!$this->_table)
-		{
-			$identifier 		= clone $this->_identifier;
-			$identifier->name	= KInflector::tableize($identifier->name);
-			$identifier->path	= array('database', 'table');
+     * Method to get a table object
+     * 
+     * Function catches KDatabaseTableExceptions that are thrown for tables that 
+     * don't exist. If no table object can be created the function will return FALSE.
+     *
+     * @return KDatabaseTableAbstract
+     */
+    public function getTable()
+    {
+        if($this->_table !== false)
+        {
+            if(!($this->_table instanceof KDatabaseTableAbstract))
+		    {   		        
+		        //Make sure we have a table identifier
+		        if(!($this->_table instanceof KServiceIdentifier)) {
+		            $this->setTable($this->_table);
+			    }
+		        
+		        try {
+		            $this->_table = $this->getService($this->_table);
+                } catch (KDatabaseTableException $e) {
+                    $this->_table = false;
+                }
+            }
+        }
 
-			$this->_table = KFactory::get($identifier);
-		}
-
-		return $this->_table;
-	}
-
+        return $this->_table;
+    }
+	
 	/**
 	 * Method to set a table object attached to the rowset
 	 *
-	 * @param	mixed	An object that implements KObjectIdentifiable, an object that
-	 *                  implements KIndentifierInterface or valid identifier string
+	 * @param	mixed	An object that implements KObjectServiceable, KServiceIdentifier object 
+	 * 					or valid identifier string
 	 * @throws	KDatabaseRowException	If the identifier is not a table identifier
 	 * @return	KDatabaseRowsetAbstract
 	 */
-	public function setTable($table)
+    public function setTable($table)
 	{
 		if(!($table instanceof KDatabaseTableAbstract))
 		{
-			$identifier = KFactory::identify($table);
-
-			if($identifier->path[0] != 'table') {
-				throw new KModelException('Identifier: '.$identifier.' is not a table identifier');
+			if(is_string($table) && strpos($table, '.') === false ) 
+		    {
+		        $identifier         = clone $this->getIdentifier();
+		        $identifier->path   = array('database', 'table');
+		        $identifier->name   = KInflector::tableize($table);
+		    }
+		    else  $identifier = $this->getIdentifier($table);
+		    
+			if($identifier->path[1] != 'table') {
+				throw new KDatabaseRowsetException('Identifier: '.$identifier.' is not a table identifier');
 			}
 
-			$table = KFactory::get($identifier);
+			$table = $identifier;
 		}
 
 		$this->_table = $table;
+
 		return $this;
+	}
+	
+	/**
+	 * Test the connected status of the row.
+	 *
+	 * @return	boolean	Returns TRUE if we have a reference to a live KDatabaseTableAbstract object.
+	 */
+    public function isConnected()
+	{
+	    return (bool) $this->getTable();
 	}
 
 	/**
 	 * Load the row from the database using the data in the row
 	 *
-	 * @return boolean	If successfull return TRUE, otherwise FALSE
+	 * @return object	If successfull returns the row object, otherwise NULL
 	 */
 	public function load()
 	{
-		$result = false;
+		$result = null;
 		
 		if($this->_new)
 		{
-            $table = $this->getTable();
+            if($this->isConnected())
+            {
+                $data  = $this->getTable()->filter($this->getData(true), true);
+		        $row   = $this->getTable()->select($data, KDatabase::FETCH_ROW);
 
-		    $data  = $table->filter($this->getData(true), true);
-		    $row   = $table->select($data, KDatabase::FETCH_ROW);
-
-		    // Set the data if the row was loaded succesfully.
-		    if(!$row->isNew())
-		    {
-			    $this->setData($row->toArray(), false);
-			    $this->_modified = array();
-			    $this->_new      = false;
+		        // Set the data if the row was loaded succesfully.
+		        if(!$row->isNew())
+		        {
+			        $this->setData($row->toArray(), false);
+			        $this->_modified = array();
+			        $this->_new      = false;
 			    
-			    $this->setStatus(KDatabase::STATUS_LOADED);
-			    $result = true;
-		    }
-		    else $this->setStatus(KDatabase::STATUS_FAILED);
+			        $this->setStatus(KDatabase::STATUS_LOADED);
+			        $result = $this;
+		        }
+            }
 		}
 	
 		return $result;
@@ -144,11 +177,24 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	 */
 	public function save()
 	{
-	    if($this->_new) {
-		    $result = $this->getTable()->insert($this);;
-		} else {
-			$result = $this->getTable()->update($this);
-		}
+	    $result = false;
+	    
+	    if($this->isConnected())
+	    {  
+	        if($this->_new) {
+	            $result = $this->getTable()->insert($this);
+		    } else {
+		        $result = $this->getTable()->update($this); 
+		    }
+	  
+	        if($result !== false) 
+	        {
+	            // Filter out any extra columns.
+	            if(((integer) $result) > 0) {       
+                    $this->_modified = array();
+	            } 
+            }
+	    }
 
 		return (bool) $result;
     }
@@ -161,9 +207,20 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	public function delete()
 	{
 		$result = false;
-
-		if(!$this->_new) {
-		    $result = $this->getTable()->delete($this);
+		
+		if($this->isConnected())
+		{
+            if(!$this->_new) 
+		    {
+		        $result = $this->getTable()->delete($this);
+		    
+		        if($result !== false) 
+	            {
+	                if(((integer) $result) > 0) {   
+	                    $this->_new = true;
+	                } 
+                }
+		    }
 		}
 
 		return (bool) $result;
@@ -176,11 +233,13 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	 */
 	public function reset()
 	{
-		$result = false;
-	    
-	    if($this->_data = $this->getTable()->getDefaults()) {
-		    $this->setStatus(null);
-		    $result = true;
+		$result = parent::reset();
+		
+		if($this->isConnected())
+		{
+	        if($this->_data = $this->getTable()->getDefaults()) {
+		        $result = true;
+		    }
 		}
 		
 		return $result;
@@ -193,10 +252,15 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	 */
 	public function count()
 	{
-		$data  = $this->getTable()->filter($this->getData(true), true);
-		$count =  $this->getTable()->count($data);
+		$result = false;
+	    
+	    if($this->isConnected())
+		{
+	        $data   = $this->getTable()->filter($this->getData(true), true);
+		    $result = $this->getTable()->count($data);
+		}
 
-		return $count;
+		return $result;
 	}
 
 	/**
@@ -210,12 +274,15 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	 */
 	public function __unset($column)
 	{
-		$field = $this->getTable()->getColumn($column);
+		if($this->isConnected())
+		{
+	        $field = $this->getTable()->getColumn($column);
 
-		if(isset($field) && $field->required) {
-			$this->_data[$column] = $field->default;
-		} else {
-			parent::__unset($column);
+		    if(isset($field) && $field->required) {
+			    $this->_data[$column] = $field->default;
+		    } else {
+			    parent::__unset($column);
+		    }
 		}
 	}
 
@@ -233,14 +300,31 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	 */
 	public function __call($method, array $arguments)
 	{ 
-	    // If the method hasn't been mixed yet, load all the behaviors.
-		if(!isset($this->_mixed_methods[$method]))
+	    if($this->isConnected())
 		{
-			foreach($this->getTable()->getBehaviors() as $behavior) {
-				$this->mixin($behavior);
-			}
+		    $parts = KInflector::explode($method);
+		    
+		     //Check if a behavior is mixed
+		    if($parts[0] == 'is' && isset($parts[1]))
+		    {
+		        if(!isset($this->_mixed_methods[$method]))
+                { 
+		             //Lazy mix behaviors
+		            $behavior = strtolower($parts[1]);
+		        
+                    if($this->getTable()->hasBehavior($behavior)) 
+                    {
+                        $this->mixin($this->getTable()->getBehavior($behavior));
+                        return true;
+		            }
+		    
+			        return false;
+                }
+		       
+                return true;
+		    }
 		}
-
+		   
 		return parent::__call($method, $arguments);
 	}
 }
